Getting the Dev Env object by name functionality moved to the DevEnvSetup class, because it can be used on both DevEnvLocalSetup and DevEnvOrgSetup.
