+++ dem/cli/command/pull_cmd.py
@@ -56,7 +56,7 @@ def execute(dev_env_name: str) -> None:
     #Get the organization's Dev Env if available.
     dev_env_org_json_deserialized = data_management.read_deserialized_dev_env_org_json()
     dev_env_org_setup = dev_env_setup.DevEnvOrgSetup(dev_env_org_json_deserialized)
-    dev_env_org = dev_env_org_setup.get_dev_env(dev_env_name)
+    dev_env_org = dev_env_org_setup.get_dev_env_by_name(dev_env_name)
     if dev_env_org is None:
         stderr.print("[red]Error: The input Development Environment is not available for the organization.[/]")
         return
