+++ tests/cli/test_create_cmd.py
@@ -9,7 +9,7 @@ import dem.cli.command.create_cmd as create_cmd
 from typer.testing import CliRunner
 from unittest.mock import patch, MagicMock, call
 
-from dem.core.dev_env_setup import DevEnv, DevEnvLocal, DevEnvLocalSetup
+from dem.core.dev_env_setup import DevEnv
 
 ## Global test variables
 
@@ -18,35 +18,6 @@ runner = CliRunner(mix_stderr=False)
 
 ## Test helpers
 ## Test cases
-def test_dev_env_name_check_match():
-    # Test setup
-    test_name = "dev_env_name"
-    expected_dev_env = MagicMock()
-    expected_dev_env.name = test_name
-    dev_env_local_setup = MagicMock()
-    dev_env_local_setup.dev_envs = []
-    dev_env_local_setup.dev_envs.append(expected_dev_env)
-
-    # Run unit under test
-    actual_dev_env = create_cmd.dev_env_name_check(dev_env_local_setup, test_name)
-
-    # Check expectations
-    assert actual_dev_env == expected_dev_env
-
-def test_dev_env_name_check_no_match():
-    # Test setup
-    test_name = "dev_env_name"
-    fake_dev_env = MagicMock()
-    fake_dev_env.name = test_name
-    dev_env_local_setup = MagicMock()
-    dev_env_local_setup.dev_envs = []
-    dev_env_local_setup.dev_envs.append(fake_dev_env)
-
-    # Run unit under test
-    actual_dev_env = create_cmd.dev_env_name_check(dev_env_local_setup, "no_matching_name")
-
-    # Check expectations
-    assert actual_dev_env is None
 
 @patch("dem.cli.command.create_cmd.container_engine.ContainerEngine")
 @patch("dem.cli.command.create_cmd.registry.list_repos")
@@ -164,18 +135,17 @@ def test_get_dev_env_descriptor_from_user(mock_get_tool_images, mock_ToolImageMe
 @patch("dem.cli.command.create_cmd.data_management.write_deserialized_dev_env_json")
 @patch("dem.cli.command.create_cmd.DevEnvLocal")
 @patch("dem.cli.command.create_cmd.get_dev_env_descriptor_from_user")
-@patch("dem.cli.command.create_cmd.dev_env_name_check")
 @patch("dem.cli.command.create_cmd.DevEnvLocalSetup")
 @patch("dem.cli.command.create_cmd.data_management.read_deserialized_dev_env_json")
 def test_execute_dev_env_creation(mock_read_deserialized_dev_env_json, mock_DevEnvLocalSetup,
-                                  mock_dev_env_name_check, mock_get_dev_env_descriptor_from_user,
+                                  mock_get_dev_env_descriptor_from_user,
                                   mock_DevEnvLocal, mock_write_deserialized_dev_env_json):
     # Test setup
     fake_deserialized_local_dev_env = MagicMock()
     mock_read_deserialized_dev_env_json.return_value = fake_deserialized_local_dev_env
     fake_dev_env_local_setup = MagicMock()
     mock_DevEnvLocalSetup.return_value = fake_dev_env_local_setup
-    mock_dev_env_name_check.return_value = None
+    fake_dev_env_local_setup.get_dev_env_by_name.return_value = None
     fake_dev_env_descriptor = MagicMock()
     mock_get_dev_env_descriptor_from_user.return_value = fake_dev_env_descriptor
     fake_dev_env_local_setup.get_deserialized.return_value = fake_deserialized_local_dev_env
@@ -189,7 +159,7 @@ def test_execute_dev_env_creation(mock_read_deserialized_dev_env_json, mock_DevE
 
     mock_read_deserialized_dev_env_json.assert_called_once()
     mock_DevEnvLocalSetup.assert_called_once_with(fake_deserialized_local_dev_env)
-    mock_dev_env_name_check.assert_called_once_with(fake_dev_env_local_setup, expected_dev_env_name)
+    fake_dev_env_local_setup.get_dev_env_by_name.assert_called_once_with(expected_dev_env_name)
     mock_get_dev_env_descriptor_from_user.assert_called_once_with(expected_dev_env_name)
     mock_DevEnvLocal.assert_called_once_with(fake_dev_env_descriptor)
     fake_dev_env_local_setup.get_deserialized.assert_called_once()
@@ -198,11 +168,10 @@ def test_execute_dev_env_creation(mock_read_deserialized_dev_env_json, mock_DevE
 @patch("dem.cli.command.create_cmd.data_management.write_deserialized_dev_env_json")
 @patch("dem.cli.command.create_cmd.get_dev_env_descriptor_from_user")
 @patch("dem.cli.command.create_cmd.typer.confirm")
-@patch("dem.cli.command.create_cmd.dev_env_name_check")
 @patch("dem.cli.command.create_cmd.DevEnvLocalSetup")
 @patch("dem.cli.command.create_cmd.data_management.read_deserialized_dev_env_json")
 def test_execute_dev_env_overwrite(mock_read_deserialized_dev_env_json, mock_DevEnvLocalSetup,
-                                   mock_dev_env_name_check, mock_confirm,
+                                   mock_confirm,
                                    mock_get_dev_env_descriptor_from_user,
                                    mock_write_deserialized_dev_env_json):
     # Test setup
@@ -211,7 +180,7 @@ def test_execute_dev_env_overwrite(mock_read_deserialized_dev_env_json, mock_Dev
     fake_dev_env_local_setup = MagicMock()
     mock_DevEnvLocalSetup.return_value = fake_dev_env_local_setup
     fake_dev_env_original = MagicMock()
-    mock_dev_env_name_check.return_value = fake_dev_env_original
+    fake_dev_env_local_setup.get_dev_env_by_name.return_value = fake_dev_env_original
     fake_tools = MagicMock()
     fake_dev_env_descriptor = {
         "tools": fake_tools
@@ -228,7 +197,7 @@ def test_execute_dev_env_overwrite(mock_read_deserialized_dev_env_json, mock_Dev
 
     mock_read_deserialized_dev_env_json.assert_called_once()
     mock_DevEnvLocalSetup.assert_called_once_with(fake_deserialized_local_dev_env)
-    mock_dev_env_name_check.assert_called_once_with(fake_dev_env_local_setup, expected_dev_env_name)
+    fake_dev_env_local_setup.get_dev_env_by_name.assert_called_once_with(expected_dev_env_name)
     mock_confirm.assert_called_once_with("The input name is already used by a Development Environment. Overwrite it?",
                                          abort=True)
     mock_get_dev_env_descriptor_from_user.assert_called_once_with(expected_dev_env_name)
@@ -238,19 +207,17 @@ def test_execute_dev_env_overwrite(mock_read_deserialized_dev_env_json, mock_Dev
 
 @patch("dem.cli.command.create_cmd.get_dev_env_descriptor_from_user")
 @patch("dem.cli.command.create_cmd.typer.confirm")
-@patch("dem.cli.command.create_cmd.dev_env_name_check")
 @patch("dem.cli.command.create_cmd.DevEnvLocalSetup")
 @patch("dem.cli.command.create_cmd.data_management.read_deserialized_dev_env_json")
 def test_execute_abort(mock_read_deserialized_dev_env_json, mock_DevEnvLocalSetup, 
-                       mock_dev_env_name_check, mock_confirm, 
-                       mock_get_dev_env_descriptor_from_user):
+                       mock_confirm, mock_get_dev_env_descriptor_from_user):
     # Test setup
     fake_deserialized_local_dev_env = MagicMock()
     mock_read_deserialized_dev_env_json.return_value = fake_deserialized_local_dev_env
     fake_dev_env_local_setup = MagicMock()
     mock_DevEnvLocalSetup.return_value = fake_dev_env_local_setup
     fake_dev_env_original = MagicMock()
-    mock_dev_env_name_check.return_value = fake_dev_env_original
+    fake_dev_env_local_setup.get_dev_env_by_name.return_value = fake_dev_env_original
     mock_confirm.side_effect = Exception()
     expected_dev_env_name = "test_dev_env"
 
@@ -260,7 +227,7 @@ def test_execute_abort(mock_read_deserialized_dev_env_json, mock_DevEnvLocalSetu
     # Check expectations
     mock_read_deserialized_dev_env_json.assert_called_once()
     mock_DevEnvLocalSetup.assert_called_once_with(fake_deserialized_local_dev_env)
-    mock_dev_env_name_check.assert_called_once_with(fake_dev_env_local_setup, expected_dev_env_name)
+    fake_dev_env_local_setup.get_dev_env_by_name.assert_called_once_with(expected_dev_env_name)
     mock_confirm.assert_called_once_with("The input name is already used by a Development Environment. Overwrite it?",
                                          abort=True)
     mock_get_dev_env_descriptor_from_user.assert_not_called()
